Tabs and grouping
Add tabs to sort notes by date ranges. Add grouping of notes by the same dates. Rename the toolbar class to header. Assign the toolbar class to the expense list. Add a regular expression to the amount input to determine if there is a floating point. Add an action button to add notes to favorites. Add a modal window to confirm the deletion of a note. Implement sorting via constants